python-matcher/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
opencv-python-headless==4.10.0.84
numpy==2.1.1
pydantic==2.8.2
Pillow==10.4.0
